# sauron-backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.Part:
    description: Detailed information about a firearm part including compatibility
      and specifications
    properties:
      created_at:
        description: Creation timestamp
        type: string
      description:
        description: Description of the part
        example: A 30-round 5.56x45 NATO polymer magazine for AR-15 rifles.
        type: string
      dimensions:
        description: Dimensions of the part
        example: 5 x 3 x 2 in
        type: string
      id:
        description: Unique identifier for the part
        example: 1
        type: integer
      images:
        description: Image URLs for the part
        example:
        - '["https://example.com/images/parts/Standard-Charging-Handle-(AR-15).jpg"]'
        items:
          type: string
        type: array
      is_prebuilt:
        description: Whether this is a pre-built component
        example: false
        type: boolean
      manufacturer_id:
        description: Reference to the manufacturer
        example: 1
        type: integer
      name:
        description: Name of the part
        example: PMAG 30 AR/M4 GEN M3
        type: string
      part_category:
        $ref: '#/definitions/models.PartCategory'
      part_category_id:
        description: Reference to the part category
        example: 18
        type: integer
      updated_at:
        description: Last update timestamp
        type: string
      weight:
        description: Weight of the part in pounds
        example: 0.54
        type: number
    type: object
  models.PartCategory:
    description: Hierarchical structure of part categories
    properties:
      child_categories:
        description: Child categories (reverse relationship)
        items:
          $ref: '#/definitions/models.PartCategory'
        type: array
      created_at:
        description: Creation timestamp
        type: string
      description:
        description: Description of the category
        example: Core upper receiver components
        type: string
      id:
        description: Unique identifier for the category
        example: 1
        type: integer
      name:
        description: Name of the category
        example: Upper Assembly
        type: string
      parent_category:
        allOf:
        - $ref: '#/definitions/models.PartCategory'
        description: Parent category (self-referential relationship)
      parent_category_id:
        description: Parent category ID for hierarchical relationships (null for top-level
          categories)
        example: 0
        type: integer
      updated_at:
        description: Last update timestamp
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@sauron.io
    name: Sauron API Support
    url: https://sauron.io/support
  description: A comprehensive API for firearms database management including parts,
    models, categories, product listings, and compatibility.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sauron.io/terms/
  title: Sauron Backend API
  version: "2.0"
paths:
...
  /parts:
    get:
      consumes:
      - application/json
      description: Get all parts with optional filtering by category, subcategory,
        or category_id
      parameters:
      - description: Filter by category name
        in: query
        name: category
        type: string
      - description: Filter by subcategory name
        in: query
        name: subcategory
        type: string
      - description: Filter by part category ID (new schema)
        in: query
        name: category_id
        type: integer
      - description: Filter by prebuilt status
        in: query
        name: is_prebuilt
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Part'
            type: array
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all parts
      tags:
      - Parts
    post:
      consumes:
      - application/json
      description: Create a new part in the database
      parameters:
      - description: Part object to create
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.Part'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Part'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a part
      tags:
      - Parts
  /parts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific part
      parameters:
      - description: Part ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Part not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a part
      tags:
      - Parts
    get:
      consumes:
      - application/json
      description: Get a specific part by its ID
      parameters:
      - description: Part ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Part'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Part not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get part by ID
      tags:
      - Parts
    put:
      consumes:
      - application/json
      description: Update an existing part
      parameters:
      - description: Part ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated part object
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.Part'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Part'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Part not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a part
      tags:
      - Parts
  /parts/{id}/compatible:
    get:
      consumes:
      - application/json
      description: Get all parts compatible with a specific part
      parameters:
      - description: Part ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Part'
            type: array
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Part not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get compatible parts
      tags:
      - Parts
      - Compatibility
  /parts/category/{category}:
    get:
      consumes:
      - application/json
      description: Get all parts in a specific category
      parameters:
      - description: Category name
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Part'
            type: array
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get parts by category
      tags:
      - Parts
...
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"